Add files via upload
the first page is for volunteers to choose a date then the items at the bottom will be different volunteer opportunities available on the date chosen. The second is for admins to be able to add volunteers and keep track of what activities they are most interested in volunteering for.